Add url to get models, change long lat type

+++ ModelX.Logic/CQRS/Models/Commands/PutModels/PutModelsModelDto.cs
@@ -7,7 +7,7 @@ public record PutModelsModelDto : IMapFrom<Model>
 {
     public int AttachmentId { get; set; }
 
-    public int Latitude { get; set; }
+    public double Latitude { get; set; }
 
-    public int Longitude { get; set; }
+    public double Longitude { get; set; }
 }
